refactor: renamed the fixedDimensions.js file
