+++ src/modules/layout/layoutHasDynamicDimension.js
@@ -1,6 +1,6 @@
 import { layoutGetAllDimensions } from './layoutGetAllDimensions'
 import { dimensionGetId } from './dimensionGetId'
-import { getPredefinedDimensions } from '../fixedDimensions'
+import { getPredefinedDimensions } from '../predefinedDimensions'
 
 export const layoutHasDynamicDimension = layout => {
     const fixedIds = Object.keys(getPredefinedDimensions())
